requirements.txt
torch>=1.9.0
torchvision>=0.10.0
pandas>=1.3.0
Pillow>=8.3.0
openpyxl>=3.0.0